pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "adl_pdf_ingest_phd"
version = "0.1.0"
description = "PDF ingestion pipeline for research paper corpus - extracts, indexes, and searches academic PDFs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ammons Data Labs"}
]
keywords = ["pdf", "elasticsearch", "postgres", "ingestion", "search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "PyMuPDF>=1.26.0",
    "psycopg2-binary>=2.9.0",
    "elasticsearch>=8.0.0,<9.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "reportlab>=4.0.0",
]

[project.scripts]
pdf-ingest = "pdf_ingest.cli:main"

[project.urls]
Repository = "https://github.com/ammons-datalabs/adl-pdf-ingest-phd"

[tool.setuptools.packages.find]
include = ["pdf_ingest*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v"
markers = [
    "integration: tests that require PostgreSQL and Elasticsearch services",
]